Add monitor shutdown func
